Add support for shared memory.
Important notes: declaring the opacity hint's status variable as volatile is
necessary. Otherwise the compiler will assume the variable remains unchanged
while the non-master chains wait for the opacity grid to be written. Also, the
freemem_opacity function is not currently called. It should be, once it is
fixed to avoid freeing null pointers.