Ignore warnings related to ignoring function results and failed inlining.
All of the warnings that were disabled by -Wno-unused-result related to the
use of fread(). We don't want to check the status of fread() every time it is
used, despite the fact that such is generally a good idea.

All of the warnings disabled by -Wno-inline related to functions that were
declared as inline, but could not be inlined because the function body exists
in a file other than the header. If the speed of inlining is really important
for any particular function, the body should be moved to its header.

Reference #2.