Set platform guard on error call
On linux gnu provides an error call in the std library implementation.
This is used to make a nice error message if memory allocation fails.
However on other platforms this is most likely not present. This
commit puts the call behind a platform macro. On non linux systems
a malloc error should just crash out with the systems normal error.
This resolves a build problem where the error symbols were not
present in the build.